refactor triple colon class after some tipple
